Refactor data management into hooks and components
