+++ src/badgeDefs.js
@@ -1,0 +1,19 @@
+export const BADGE_DEFS = [
+  { id: 'eeg', title: 'EEG' },
+  { id: 'eeg2', title: 'EEG2' },
+  { id: 'experiment', title: 'Experiment' },
+  { id: 'facereader', title: 'Facereader' },
+  { id: 'excursie', title: 'Excursie' },
+  { id: 'groupname', title: 'Groepsnaam & mascotte' },
+  { id: 'homework', title: 'Homework' },
+  { id: 'kennistoets', title: 'Kennistoets' },
+  { id: 'leeswerk', title: 'Leeswerk' },
+  { id: 'lunch', title: 'Lunch' },
+  { id: 'meeting', title: 'Meeting with commissioner' },
+  { id: 'minorbehaald', title: 'Minor behaald' },
+  { id: 'namen', title: 'Namen badge' },
+  { id: 'partycommittee', title: 'Party committee' },
+  { id: 'project', title: 'Project' },
+  { id: 'pubquiz', title: 'Pubquiz' },
+  { id: 'pupil-labs', title: 'Pupil labs' },
+];
